+++ CHANGES.rst
@@ -4,6 +4,7 @@ Changelog
 2.4.0 (unreleased)
 ------------------
 
+- #2255 Fix attachments from retracted or rejected analyses are not ignored
 - #2201 Allow manual selection of units on results entry
 - #2258 Reduce conflict errors on number generation
 - #2256 Do not keep DX UID reference field back-references per default